api/requirements.txt
#dependencies
fastapi 
uvicorn[standard] 
requests 
sqlalchemy 
psycopg2-binary 
PyYAML 
celery[redis] 
tenacity 
pytest 
pyyaml 
python-jose[cryptography]
passlib[bcrypt]
python-multipart
asyncpg
pydantic